+++ src/sympc/protocol/beaver/beaver.py
@@ -18,10 +18,10 @@ from typing import Tuple
 import torch
 import torchcsprng as csprng  # type: ignore
 
-from sympc.exceptions import EmptyPrimitiveStore
 from sympc.store import register_primitive_generator
 from sympc.store import register_primitive_store_add
 from sympc.store import register_primitive_store_get
+from sympc.store.exceptions import EmptyPrimitiveStore
 from sympc.tensor import MPCTensor
 from sympc.tensor import ShareTensor
 from sympc.utils import count_wraps
