+++ src/sympc/store/exceptions.py
